// src/app/subModulos/interfaces/ubicacion/local-ubicacion-interface.ts
/*export interface MyHomeBrewery {
    taps: Array<Taps>;    
    barrels: Array<Barrels>;
}*/

export interface LocalInterface {

    iLocalId : string;
    cLocalDescripcion :string;
}

/*export interface Barrels {
    id : number;
    name: string;
    width: number;
    height: number;
    quantity: number;
}*/
  